texture error 1

+++ MGL/include/glm_params.h
@@ -193,8 +193,8 @@ typedef struct GLMParams_t {
     GLuint map1_grid_segments;
     GLuint map2_grid_domain;
     GLuint map2_grid_segments;
-    GLuint polygon_offset_units;
-    GLuint polygon_offset_factor;
+    GLfloat polygon_offset_units;
+    GLfloat polygon_offset_factor;
     GLuint texture_binding_1d;
     GLuint texture_binding_2d;
     GLuint client_attrib_stack_depth;
